refactor(qr): give reader preparation, diagnostics, acquisition and the pump their own modules
@/qr/decode keeps its exact public surface: tests/ui/helpers/module-mocks.ts
mocks that path and supplies only startQrScan.
